rename polyon mapping from quattroshapes->boundary to be more generic/reusable, deleted redundant file. Updated tests
